added error handling for decimal points in input

+++ app/run_create_prime_table.rb
@@ -1,3 +1,3 @@
 require_relative 'prime_table'
 
-PrimeTable.new.create_table(ARGV[0].to_i)
+PrimeTable.new.create_table(ARGV[0].to_f)
